Make Examples tests more robust
- Remove navigation animations
- Fix potential race condition with finish notifications

[run app device tests]
