+++ Apps/Examples/Examples/Controllers/ExampleTableViewController.swift
@@ -88,12 +88,8 @@ extension ExampleTableViewController {
           example = allExamples[indexPath.row]
         }
 
-        show(example: example)
-    }
-
-    internal func show(example: Example) {
-        let exampleToDisplay = makeViewController(for: example)
-        navigationController?.pushViewController(exampleToDisplay, animated: true)
+        let exampleViewController = example.makeViewController()
+        navigationController?.pushViewController(exampleViewController, animated: true)
     }
 
     public func filterContentForSearchText(_ searchText: String) {
@@ -102,49 +98,5 @@ extension ExampleTableViewController {
         }
 
       tableView.reloadData()
-    }
-
-    public func makeViewController(for example: Example) -> UIViewController {
-        guard let exampleClass = example.type as? UIViewController.Type else {
-            fatalError("Unable to get class name from example named \(example.type)")
-        }
-
-        let exampleViewController = exampleClass.init()
-        exampleViewController.title = example.title
-        exampleViewController.navigationItem.largeTitleDisplayMode = .never
-
-        let association = ExampleAssociation(example: example, viewController: exampleViewController)
-        objc_setAssociatedObject(exampleViewController, &ExampleAssociationHandle, association, objc_AssociationPolicy.OBJC_ASSOCIATION_RETAIN_NONATOMIC)
-        let barButtonItem = UIBarButtonItem(title: "Info", style: .plain, target: association, action: #selector(ExampleAssociation.presentAlert))
-
-        exampleViewController.navigationItem.setRightBarButton(barButtonItem, animated: false)
-
-        return exampleViewController
-    }
-}
-
-private var ExampleAssociationHandle: UInt8 = 0
-internal class ExampleAssociation: NSObject {
-
-    internal let example: Example
-    internal weak var viewController: UIViewController?
-
-    internal init(example: Example, viewController: UIViewController) {
-        self.example = example
-        self.viewController = viewController
-        super.init()
-    }
-
-    @objc func presentAlert() {
-        guard let viewController = viewController else {
-            return
-        }
-
-        let alert = UIAlertController(title: "About this example",
-                                      message: example.description,
-                                      preferredStyle: .alert)
-        let action = UIAlertAction(title: "Got it", style: .default, handler: nil)
-        alert.addAction(action)
-        viewController.present(alert, animated: true, completion: nil)
     }
 }
